+++ CHANGELOG.md
@@ -2,8 +2,24 @@
 
 ## Unreleased
 
+- fix(nuxt): Remove unused parameter in sentry-example-api template ([#734](https://github.com/getsentry/sentry-wizard/pull/734))
+
+## 3.36.0
+
+- Remove Profiling for Android ([#720](https://github.com/getsentry/sentry-wizard/pull/720))
+- Add downgrade path to nitro 2.9.7 ([#725](https://github.com/getsentry/sentry-wizard/pull/725))
+
+## 3.35.0
+
 - feat: Pin JS SDK versions to v8 (#712)
 - Remove enableTracing for Cocoa ([#715](https://github.com/getsentry/sentry-wizard/pull/715))
+- feat(nuxt): Add nuxt wizard ([#719](https://github.com/getsentry/sentry-wizard/pull/719))
+ 
+Set up the Sentry Nuxt SDK in your app with one command:
+
+```sh
+npx @sentry/wizard@latest -i nuxt
+```
 
 ## 3.34.4